requirements.txt
scikit-learn
scipy
plotly
kaleido
ipykernel
numpy
matplotlib
pandas
tqdm
nbformat
tensorboard
torch
torchvision
torchaudio
wandb
MulticoreTSNE
optuna
hydra-core
python-dotenv
lightgbm
synthcity
pytrial
transformers
datasets
accelerate
evaluate
einops
# rouge